requirements.txt
pyperclip==1.8.2
qrcode==7.3
Pillow==8.3.1